refactor(pool-page): extract _tabs/lps-tab
Move LpsTab (~280 lines) and its private helper
isLiquidityPositionSchemaError out of page.tsx into _tabs/lps-tab.tsx.
The helper stays a module-private function since it's only consumed by
LpsTab's error branch.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green; lazy-mount loop in __tests__/page.test.tsx
  re-runs every iteration, confirming PoolLpPositions still only fires
  on tab=providers)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=providers vs baseline:
  zero structural changes, console clean

page.tsx: 1540 → 1257 lines (-283).
